refactor: transaction-service Module generate and refactor
