+++ transaction-service/build.gradle
@@ -1,15 +1,24 @@
 plugins {
     id 'java'
-}
-
-group 'com.whalewatch'
-version '0.0.1-SNAPSHOT'
-
-repositories {
-    mavenCentral()
+    id 'org.springframework.boot' version '3.3.4'
+    id 'io.spring.dependency-management'
 }
 
 dependencies {
+    implementation project(':common')
+    implementation project(':service')
+
+    // Spring Boot, JPA
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+
+    // MapStruct
+    implementation 'org.mapstruct:mapstruct:1.5.5.Final'
+    annotationProcessor 'org.mapstruct:mapstruct-processor:1.5.5.Final'
+
+    //Kafka
+    implementation 'org.springframework.kafka:spring-kafka'
+
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
 }
